Add extra guards to complex type handlers
Complex types for CefValue such as Binary, Dictionary and List can take
a different owner (eg: a Dictionary may be assigned to an element of a
List), which turns the handle invalid. Ideally we should discard the
handles for such elements as quickly as possible, but in practice the
wrapper objects in .NET can stay in memory for a long time before the GC
decides to run their finalizers. By running the finalizers late, the
probability that the handle has become invalid is high.

In a previous commit, we wrapped most usages of complex values within
`using` blocks, calling the Dispose immediately instead of the finalizer
a lot later, which reduced incidence of crashes. Apparently that was not
enough, so with this commit we add a few more protection on remaining
usages of Lists, Dictionaries and Binaries.