memleak: break object reference cycle so GC works
The GC is smart enough to deal with reference cycles. But in this
particular case the GC can't work its magic, because the referential
cycle has a part in the native land, to which the GC has no visibility.

This cycle works with something like:

  CefBrowser -> [ native ] -> CefClient -> (...) -> CefBrowser

CefClient is kept active because there's a native reference to it, but
the GC doesn't know what a native reference is... all it knows is that
the object was pinned (we must pin it to keep the native reference
valid). So the GC only sees something like

  (pinned root) -> CefClient -> (...) -> CefBrowser

This prevents the GC from collecting CefBrowser, which prevents the
native reference from being cleared, which prevents CefClient to be
cleared.

We break this cycle by making objects along the chain (represented above
by "(...)") clear their reference to CefBrowser upon disposal.